First feature-complete'ish version
This is a port of the Perl script scope that I used for many years now
to create tags databases.

It has all features of the Perl script. The design and implementation
are quick and dirty, I wanted to have something usable first. It also
misses tests, all of them.