Get rid of separate thread to crawl files
Restructure the program so that the MIME worker threads push their file
names for tagging into the pipe(2) for ctags and cscope themselves. That
frees up the main thread and allows us to get rid of the separate
crawler thread. The main thread now crawls all files itself.

That simplifies the pipeline.